Updated/Working User and Auth routes
