+++ BDR_API_Web_Proj/src/main/java/ch/heigvd/aliment/alimentController.java
@@ -1,0 +1,18 @@
+package ch.heigvd.aliment;
+
+import io.javalin.http.Context;
+
+public class alimentController {
+    public static void getMany(Context ctx) {
+
+    }
+
+    public static void create(Context context) {
+    }
+
+    public static void update(Context context) {
+    }
+
+    public static void delete(Context context) {
+    }
+}
