+++ BDR_API_Web_Proj/src/main/java/ch/heigvd/auth/AuthController.java
@@ -1,47 +1,37 @@
 package ch.heigvd.auth;
 
-import ch.heigvd.users.User;
-import ch.heigvd.users.UserLogin;
+import ch.heigvd.user.User;
 import io.javalin.http.*;
 import java.sql.*;
 
 public class AuthController {
 
-    private final String jdbcUrl;
-    private final String dbUsername;
-    private final String dbPassword;
-
-    public AuthController(String jdbcUrl, String dbUsername, String dbPassword) {
-        this.jdbcUrl = jdbcUrl;
-        this.dbUsername = dbUsername;
-        this.dbPassword = dbPassword;
-    }
-
-    private Connection getConnection() throws SQLException {
-        return DriverManager.getConnection(jdbcUrl, dbUsername, dbPassword);
+    private Connection conn;
+    public AuthController(Connection connection) {
+        conn = connection;
     }
 
     public void login(Context ctx) throws SQLException {
 
-        UserLogin loginUser = ctx.bodyValidator(UserLogin.class)
+        User loginUser = ctx.bodyValidator(User.class)
                 .check(obj -> obj.email != null, "Missing email")
                 .check(obj -> obj.password != null, "Missing password")
                 .get();
 
         System.out.println("User want to login with email="+loginUser.email+" and password="+loginUser.password+".");
 
-        try (Connection conn = getConnection()) {
-            PreparedStatement stmt = conn.prepareStatement("SELECT * FROM utilisateur WHERE email = ? AND motdepasse = ?"); // Note: Storing and comparing passwords in plain text is not secure
-            stmt.setString(1, loginUser.email);
-            stmt.setString(2, loginUser.password);
 
-            ResultSet rs = stmt.executeQuery();
-            if (rs.next()) {
-                ctx.cookie("user", rs.getString("email"));
-                ctx.status(HttpStatus.NO_CONTENT);
-                return;
-            }
+        PreparedStatement stmt = conn.prepareStatement("SELECT * FROM utilisateur WHERE email = ? AND motdepasse = ?"); // Note: Storing and comparing passwords in plain text is not secure
+        stmt.setString(1, loginUser.email);
+        stmt.setString(2, loginUser.password);
+
+        ResultSet rs = stmt.executeQuery();
+        if (rs.next()) {
+            ctx.cookie("user", rs.getString("email"));
+            ctx.status(HttpStatus.NO_CONTENT);
+            return;
         }
+
 
         throw new UnauthorizedResponse();
     }
@@ -58,23 +48,64 @@ public class AuthController {
             throw new UnauthorizedResponse();
         }
 
-        try (Connection conn = getConnection()) {
-            PreparedStatement stmt = conn.prepareStatement("SELECT * FROM utilisateur WHERE email = ?");
-            stmt.setString(1, userEmail);
 
-            ResultSet rs = stmt.executeQuery();
-            if (rs.next()) {
-                User user = new User();
-                user.email = rs.getString("email");
-                user.firstName = rs.getString("prenom");
-                user.lastName = rs.getString("nom");
-                user.password = rs.getString("motdepasse"); // Note: Storing passwords in plain text is not secure
+        PreparedStatement stmt = conn.prepareStatement("SELECT * FROM utilisateur WHERE email = ?");
+        stmt.setString(1, userEmail);
 
-                ctx.json(user);
-                return;
-            }
+        ResultSet rs = stmt.executeQuery();
+        if (rs.next()) {
+            User user = new User();
+            user.email = rs.getString("email");
+            user.firstName = rs.getString("prenom");
+            user.lastName = rs.getString("nom");
+            user.password = rs.getString("motdepasse"); // Note: Storing passwords in plain text is not secure
+
+            ctx.json(user);
+            return;
         }
 
         throw new UnauthorizedResponse();
+    }
+
+    public boolean validLoggedUser(Context ctx) throws SQLException {
+        String userEmail = ctx.cookie("user");
+
+        if (userEmail == null) {
+            throw new UnauthorizedResponse();
+        }
+
+        String email = ctx.pathParam("email");
+
+        if(userEmail.equalsIgnoreCase("admin")) {
+            return true;
+        }
+
+        if (userEmail.equalsIgnoreCase(email) == false)  {
+            throw new UnauthorizedResponse();
+        }
+
+
+
+        PreparedStatement stmt = conn.prepareStatement("SELECT * FROM utilisateur WHERE email = ?");
+        stmt.setString(1, userEmail);
+
+        ResultSet rs = stmt.executeQuery();
+        if (rs.next()) {
+            User user = new User();
+            user.email = rs.getString("email");
+            user.firstName = rs.getString("prenom");
+            user.lastName = rs.getString("nom");
+            user.password = rs.getString("motdepasse"); // Note: Storing passwords in plain text is not secure
+
+            if (user != null) {
+                if (user.email.isEmpty() == false) {
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+        }
+
+        return false;
     }
 }
