--- BDR_API_Web_Proj/src/main/java/ch/heigvd/users/User.java
@@ -1,13 +1,0 @@
-package ch.heigvd.users;
-
-public class User {
-
-    public String firstName;
-    public String lastName;
-    public String email;
-    public String password;
-
-    public User() {
-        // Empty constructor for serialisation/deserialization
-    }
-}
